certificate: move test in own module
